NDC_manager/requirements.txt
asgiref==3.7.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
Django==4.2.11
django-cors-headers==3.14.0
django-mysql==3.9.0
djangorestframework==3.15.0
firebase-admin==6.5.0
gcloud==0.18.3
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
jws==0.1.3
MarkupSafe==2.1.5
msgpack==1.0.8
mysql-connector-python==8.3.0
mysqlclient==2.2.4
oauth2client==4.1.3
pika==1.3.2
pip==23.0.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing== 3.1.2
Pyrebase4==4.7.1
python-jwt==2.0.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==58.1.0
six==1.16.0
sqlparse== 0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.18